Change made to settings.py DB_NAME
DB_NAME needs to be set to ppp_db if we are using
same db in docker network we are currently running on
production. so when container is run env DB_NAME
should be set to this in the event that we are
proceeding in an unchanged manner there.  Although
the idea is to change this later and use a
docker-compose.yml file that has been made but is
untracked.